Update config files (#14)
* Refactor: split Docker and entrypoint configs for dev and prod environments

+++ video_app/signals.py
@@ -3,6 +3,8 @@ from .models import Video
 from django.dispatch import receiver
 from django.db.models.signals import post_save, post_delete
 from .tasks import queue_video_processing
+from django.conf import settings
+from urllib.parse import urlparse
 import logging
 
 logger = logging.getLogger(__name__)
@@ -26,9 +28,31 @@ def video_post_save(sender, instance: Video, created, **kwargs):
 @receiver(post_delete, sender=Video)
 def auto_delete_file_on_delete(sender, instance, **kwargs):
     """
-    Deletes file from filesystem
+    Deletes file and thumbnail from filesystem
     when corresponding 'Video' object is deleted.
     """
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
+    logger.info("post_delete triggered for Video ID %s", instance.id)
+    
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        try:
             os.remove(instance.video_file.path)
+            logger.info("Deleted video file: %s", instance.video_file.path)
+        except Exception as e:
+            logger.warning(
+                "Could not delete video file %s: %s", instance.video_file.path, e)
+
+    if instance.thumbnail_url:
+        parsed = urlparse(instance.thumbnail_url)
+        thumb_path = parsed.path.replace("/media/", "", 1)
+        thumb_full_path = os.path.join(settings.MEDIA_ROOT, thumb_path)
+
+        if os.path.isfile(thumb_full_path):
+            try:
+                os.remove(thumb_full_path)
+                logger.info("Deleted thumbnail: %s", thumb_full_path)
+            except Exception as e:
+                logger.warning(
+                    "Could not delete thumbnail: %s: $s", thumb_full_path, e)
+
+        else:
+            logger.info("Thumbnail file not found: %s", thumb_full_path)
